rewrite core logic
This skips all the installation steps if the user has run this script
before and already has signing keys.

We also have a much simpler and safer loop for finding unsigned EFI
files and signing them.

Finally, we no longer use the file database, as Nobara puts version
information in the kernel filenames, and we would only be filling up a
database with filenames that no longer exist over time.