clean up output
The original script had some unprofessional humor. While I personally
quite appreciated it, it's not for everyone, and keeping a professional
tone should help the script be perceived as more trustworthy and easier
to understand for non-native English readers.

We also print the reboot message at the end only if the system is not
already in Secure Boot mode. This wasn't a concern before as the script
always performed full setup, so it could be assumed that the system was
never in Secure Boot mode at the end of the script. Now that we can run
the script after initial setup to only do key signing, it may often be
the case.